fix Componentization
